Create site.md
Think about someone who already knows WordPress taking over the running of the site itself: how do they add new content – as a page or a post? In what category? You may have used a plugin for a calendar of events or something, which needs to be described. Do not recreate WordPress documentation, but clearly explain to the client’s site maintainer how this particular site is organised and can be maintained.